client day 17

+++ controllers/users.js
@@ -26,15 +26,22 @@ router.post("/add", async (req, res) => {
 
     const record = {
       name: req.body.name,
-      email: req.body.email.toLowerCase,
+      email: req.body.email.toLowerCase(),
       phoneNumber: req.body.phoneNumber,
       password: await bcrypt.hash(req.body.password, 10),
       type: req.body.type,
       createdOn: new Date()
     }
 
-    if(req.body.type === userTypes.USER_TYPE__STANDARD)
-      record.departmentId = req.body.departmentId;
+    if(req.user.type === userTypes.USER_TYPE__STANDARD)
+      {
+        record.departmentId = req.user.departmentId;
+        record.type = userTypes.USER_TYPE__STANDARD
+      }else{
+        if(req.body.type === userTypes.USER_TYPE__STANDARD)
+          record.departmentId = req.body.departmentId;
+          record.departmentId = req.body.departmentId;
+      }
 
     const user = new User(record)
 
@@ -56,7 +63,8 @@ router.post("/edit", async (req, res) => {
 
     const user = await User.findById(req.body.id);
     if (!user) throw new Error("User does not exists");
-
+    if(req.user.type === userTypes.USER_TYPE__STANDARD && user.departmentId.toString() !== req.user.departmentId.toString())
+      throw new Error('invalid request');
 
     const record = {
       name: req.body.name,
@@ -79,7 +87,7 @@ router.post("/edit", async (req, res) => {
   }
 });
 
-router.delete("/delete", async (req, res) => {
+router.post("/delete", async (req, res) => {
     // console.log(req.body.id)
     try {
         if (!req.body.id) throw new Error("User id is required")
@@ -87,6 +95,12 @@ router.delete("/delete", async (req, res) => {
 
         const user = await User.findById(req.body.id)
         if (!user) throw new Error("User does not exists")
+
+        if(req.body.id === req.user._id.toString())
+          throw new Error("invalid request")
+
+        if(req.user.type === userTypes.USER_TYPE__STANDARD && user.departmentId.toString() !== req.user.departmentId.toString())
+        throw new Error('invalid request');
 
         await User.findOneAndDelete(req.body.id)
         res.json({ success: true })
@@ -146,7 +160,7 @@ router.post("/signin", async (req, res) => {
         if (!req.body.email) throw new Error("Email is required");
         if (!req.body.password) throw new Error("password is required")
         const user = await User.findOne({ email: req.body.email })
-        console.log(req.body)
+        
         if (!user) throw new Error("Email or password is incorrect")
         if (!(await bcrypt.compare(req.body.password, user.password)))
             throw new Error('Email or password is incorrect')
